Check return value of BIO_new_file
BIO_new_file can be failed when a file does not existed.
Even though before calling aff_bom:read_files check whether a file is
existed or not, it can be failed when file is deleted in context
switching(race condition). Therefore, it seems better to check the
return value of BIO_new_file.