Do not force include and library dir in configure
Forcing the include and library path is not good when the user
has a multiple prefix (e.g. /opt/local and /usr/local) for some reason.
This situation is typical for macOS (MacPorts and Homebrew).

By this change, user-specified ones are prefered over guessed prefixes.
This behavior is more straightforward because the user can have the control
of what header and library is to be linked.